refactor(cli): split structural verify runtime
Move spec-independent verify into its own runtime file now that the verify golden matrix covers structural PASS and FAIL behavior. Keep RunVerify dispatch and helper calls unchanged so the split stays behavior-preserving.
