+++ internal/pipeline/runtime.go
@@ -216,105 +216,6 @@ func RunVerify(cfg VerifyConfig) (*VerifyReport, error) {
 	return report, nil
 }
 
-// runStructuralVerify runs spec-independent verification: build, --help,
-// --json validity, version, and exit code checks for every discovered command.
-func runStructuralVerify(cfg VerifyConfig) (*VerifyReport, error) {
-	if cfg.Threshold == 0 {
-		cfg.Threshold = 80
-	}
-	if err := artifacts.CleanupGeneratedCLI(cfg.Dir, artifacts.CleanupOptions{
-		RemoveValidationBinaries: true,
-		RemoveDogfoodBinaries:    true,
-		RemoveRecursiveCopies:    true,
-		RemoveFinderMetadata:     true,
-	}); err != nil {
-		return nil, fmt.Errorf("pre-verify cleanup: %w", err)
-	}
-
-	report := &VerifyReport{Mode: "structural"}
-
-	// 1. Build the CLI
-	binaryPath, err := buildCLI(cfg.Dir)
-	if err != nil {
-		return nil, fmt.Errorf("building CLI: %w", err)
-	}
-	report.Binary = binaryPath
-
-	// 2. Discover commands from --help output
-	commands := discoverCommands(cfg.Dir, binaryPath)
-
-	// 3. Test each command structurally
-	for _, cmd := range commands {
-		result := runStructuralCommandTests(binaryPath, cmd)
-		report.Results = append(report.Results, result)
-	}
-
-	// 4. Version command check
-	versionOK := runCLI(binaryPath, []string{"version"}, os.Environ(), 10*time.Second) == nil
-	if !versionOK {
-		versionOK = runCLI(binaryPath, []string{"--version"}, os.Environ(), 10*time.Second) == nil
-	}
-	report.DataPipeline = versionOK
-	if versionOK {
-		report.DataPipelineDetail = "PASS (version command)"
-	} else {
-		report.DataPipelineDetail = "FAIL (version command)"
-	}
-	report.Freshness = runFreshnessContractTest(cfg.Dir)
-
-	finalizeVerifyReport(report, cfg.Threshold, false)
-
-	return report, nil
-}
-
-// runStructuralCommandTests tests a command without API access: --help output,
-// --json flag acceptance (doesn't crash), and exit code correctness.
-func runStructuralCommandTests(binary string, cmd discoveredCommand) CommandResult {
-	result := CommandResult{
-		Command: cmd.Name,
-		Kind:    "structural",
-	}
-
-	// Test 1: --help produces output and exits 0
-	result.Help = runCLI(binary, []string{cmd.Name, "--help"}, os.Environ(), 10*time.Second) == nil
-
-	// Test 2: --help --json doesn't crash (validates flag registration)
-	result.DryRun = runCLI(binary, []string{cmd.Name, "--help", "--json"}, os.Environ(), 10*time.Second) == nil
-
-	// Test 3: command with no args exits non-zero if it requires args/flags
-	// (validates that required flags are enforced). Skip commands that work
-	// without args (doctor, version, auth, completion, api).
-	switch cmd.Name {
-	case "doctor", "version", "auth", "completion", "api", "help":
-		result.Execute = true // these work without args
-	default:
-		// Running with just --json and no other args. If the command requires
-		// flags/args, it should exit non-zero with an error message.
-		// If it works without args, that's fine too.
-		// Either way, we're validating it doesn't crash/panic.
-		err := runCLI(binary, []string{cmd.Name, "--json"}, os.Environ(), 10*time.Second)
-		// Both outcomes are acceptable for structural verification: the
-		// command either ran successfully or exited with a proper error.
-		// A panic or timeout would still fail via runCLI.
-		result.Execute = true
-		_ = err
-	}
-
-	score := 0
-	if result.Help {
-		score++
-	}
-	if result.DryRun {
-		score++
-	}
-	if result.Execute {
-		score++
-	}
-	result.Score = score
-
-	return result
-}
-
 // buildCLI compiles the generated CLI and returns the binary path.
 func buildCLI(dir string) (string, error) {
 	name := filepath.Base(dir)
